Merge ScnTimestamp and Timestamp + scn storage implementations simplification
The memory and zookeeper SCN storage logic has been simplified. There is no more spinning in storage implementation with drifting clock. The storage returns an empty sequence and rely on the SCN client to retry later.

The storage never returns sequences over the maximum per current millisecond and rely again on the SCN client to retry if more are needed.

The SCN client ensure that the maximum sequence limit per millisonds is never exceeded per single callback. This limit may be exceeded when the SCN client aggregate multiple client fetch in a single server batch calls.